requirements.txt
fastapi
psycopg2-binary
six
uvicorn